+++ packages/common/index.ts
@@ -6,11 +6,4 @@ export * from './utils/sleep';
 export * from './utils/config';
 export * from './utils/unzip';
 export * from './utils/directories';
-
-export const JsonSaveParse = (str: string, errorReturn: any) => {
-    try {
-        return JSON.parse(str);
-    } catch (e) {
-        return errorReturn;
-    }
-};
+export * from './utils/json';
