+++ packages/common/utils/json.ts
@@ -1,0 +1,7 @@
+export const JsonSaveParse = (str: string, errorReturn: any) => {
+    try {
+        return JSON.parse(str);
+    } catch (e) {
+        return errorReturn;
+    }
+};
